Hotfix: fix several cast warnings. Add more vector math functions.

+++ src/rush/matrix/mat_base.h
@@ -336,16 +336,16 @@ namespace rush {
 #ifdef RUSH_GLM
 
         Mat(const glm::mat<Columns, Rows, Type>& o) {
-            for (int c = 0; c < Columns; ++c) {
-                for (int r = 0; r < Columns; ++r) {
+            for (size_t c = 0; c < Columns; ++c) {
+                for (size_t r = 0; r < Columns; ++r) {
                     operator()(c, r) = o[c][r];
                 }
             }
         }
 
         Mat(const glm::mat<Columns, Rows, Type>&& o) {
-            for (int c = 0; c < Columns; ++c) {
-                for (int r = 0; r < Columns; ++r) {
+            for (size_t c = 0; c < Columns; ++c) {
+                for (size_t r = 0; r < Columns; ++r) {
                     operator()(c, r) = std::move(o[c][r]);
                 }
             }
@@ -353,8 +353,8 @@ namespace rush {
 
         operator glm::mat<Columns, Rows, Type>() const {
             glm::mat<Columns, Rows, Type> result;
-            for (int c = 0; c < Columns; ++c) {
-                for (int r = 0; r < Columns; ++r) {
+            for (size_t c = 0; c < Columns; ++c) {
+                for (size_t r = 0; r < Columns; ++r) {
                     result[c][r] = operator()(c, r);
                 }
             }
